WIP - Break up compilerSupplier into some other classes that handle loading a compiler
