+++ packages/truffle-compile/compilerSupplier/loadingStrategies/Bundled.js
@@ -1,0 +1,15 @@
+const solc = require("solc");
+const LoadingStrategy = require("./LoadingStrategy");
+
+class Bundled extends LoadingStrategy {
+  load() {
+    return this.getBundledSolc();
+  }
+
+  getBundledSolc() {
+    this.removeListener();
+    return solc;
+  }
+}
+
+module.exports = Bundled;
