+++ packages/truffle-compile/compilerSupplier/loadingStrategies/Native.js
@@ -1,0 +1,9 @@
+const LoadingStrategy = require("./LoadingStrategy");
+
+class Native extends LoadingStrategy {
+  load() {
+    return this.getBuilt("native");
+  }
+}
+
+module.exports = Native;
