+++ packages/truffle-compile/compilerSupplier/loadingStrategies/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  Bundled: require("./Bundled"),
+  Docker: require("./Docker"),
+  Local: require("./Local"),
+  Native: require("./Native"),
+  VersionRange: require("./VersionRange")
+};
